style: add skeleton component & wrap form in suspense
